Remove types and functions from field selection component files
