+++ src/components/tables/FieldSelection/Rows.tsx
@@ -9,77 +9,13 @@ import {
 } from 'context/Theme';
 import { orderBy } from 'lodash';
 import { useBinding_currentBindingUUID } from 'stores/Binding/hooks';
-import { SortDirection, TableColumns } from 'types';
+import { basicSort_string } from 'utils/misc-utils';
 import {
-    basicSort_string,
-    compareInitialCharacterType,
-} from 'utils/misc-utils';
-import { optionalColumnIntlKeys } from '.';
-
-interface RowProps {
-    columns: TableColumns[];
-    row: CompositeProjection;
-}
-
-interface RowsProps {
-    columnToSort: string;
-    columns: TableColumns[];
-    data: CompositeProjection[];
-    sortDirection: SortDirection;
-}
-
-const compareConstraintTypes = (
-    a: CompositeProjection,
-    b: CompositeProjection,
-    ascendingSort: boolean
-): number => {
-    // If a and b have constraint types but they are not equal, compare their severity.
-    // The projection with the greater severity should appear first.
-    if (
-        a.constraint &&
-        b.constraint &&
-        a.constraint.type !== b.constraint.type
-    ) {
-        return ascendingSort
-            ? a.constraint.type - b.constraint.type
-            : b.constraint.type - a.constraint.type;
-    }
-
-    // If a and b do not have defined constraint types or their constraint types are equal,
-    // perform an ascending, alphabetic sort on their fields.
-    return ascendingSort
-        ? a.field.localeCompare(b.field)
-        : b.field.localeCompare(a.field);
-};
-
-const constraintTypeSort = (
-    a: CompositeProjection,
-    b: CompositeProjection,
-    sortDirection: SortDirection
-) => {
-    const sortResult = compareInitialCharacterType(a.field, b.field);
-
-    if (typeof sortResult === 'number') {
-        return sortResult;
-    }
-
-    // If a does not have a constraint type and b does then return >0 to put b first
-    if (!a.constraint && b.constraint) {
-        return 1;
-    }
-
-    // If a has a constraint type and b does not then return <0 to put a first
-    if (a.constraint && !b.constraint) {
-        return -1;
-    }
-
-    const ascendingSort = sortDirection === 'asc';
-
-    return compareConstraintTypes(a, b, ascendingSort);
-};
-
-const displayOptionalColumn = (columns: TableColumns[], intlKey: string) =>
-    columns.some((column) => column.headerIntlKey === intlKey);
+    constraintTypeSort,
+    displayOptionalColumn,
+    optionalColumnIntlKeys,
+} from './shared';
+import { RowProps, RowsProps } from './types';
 
 function Row({ columns, row }: RowProps) {
     const currentBindingUUID = useBinding_currentBindingUUID();
